refactor: delete spaces in modules
