+++ src/global.scss
@@ -1,6 +1,7 @@
 @import "./utils/scss/constants.scss";
 @import "./utils/scss/mixin.scss";
 @import "../src/modules/layouts/textBookChapter/textBookCharter.scss";
+
 * {
   padding: 0;
   margin: 0;
