+++ src/index.html
@@ -8,7 +8,7 @@
     <link rel="shortcut icon" href="./icon.ico">
   </head>
   <body>
-
+    
   </body>
 </html>
