+++ src/modules/authentication/authModal.scss
@@ -99,5 +99,3 @@
   text-align: center;
   color: $colorRed;
 }
-
-
