+++ src/modules/layouts/header/header.scss
@@ -63,6 +63,8 @@
 
 .navigation__item.active .menu__icon {
   transform: translate3d(-95%, 0, 0);
+
+  cursor: default;
 }
 
 .navigation__item.active::before {
@@ -72,6 +74,8 @@
 .menu__text.active {
   transform: translate3d(0, 0, 0);
   transition: transform calc(0.45s / 1.5);
+
+  cursor: default;
 }
 
 .icon {
@@ -82,20 +86,6 @@
   height: 2.3em;
   stroke-width: 10;
   stroke-miterlimit: 10;
-}
-
-.active #home-anm {
-  animation: home 1s;
-}
-
-@keyframes home {
-  25% {
-    transform: translate3d(0, -0.1em, 0);
-  }
-
-  50% {
-    transform: translate3d(0, 0.1em, 0);
-  }
 }
 
 #strategy-anm {
