+++ src/modules/layouts/miniGames/miniGames.scss
@@ -3,6 +3,11 @@
 
 .games {
   background: url('../../../assets/png/games-bg.png') 100% 100% no-repeat;
+  background-size: cover;
+  min-height: 80vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 
 .games-wrapper {
